bookmarks: Fix on_delete_range
When deleting a range that did not start at the beginning of a line,
the bookmarks contained in the range were left in the buffer, in the
middle of the line. This broke bookmark navigation. And in case the
line didn't already have a bookmark at the beginning, it caused
the bookmark pixbuf (star symbol) in the line marks to disappear.

https://bugzilla.gnome.org/show_bug.cgi?id=719345