git: Prevent trying to find a git repository for every file in a directory
Instead just cache the failure, this also means that we don't use
the weakref value dict anymore and instead rely on the window's
focus-in-event handler to clear all the cache. This allows a user
to come back into gedit and for all of the repositories to be
rediscovered, who know maybe the user ran `git init`.