git: Only return a repo for a location in the .git dir when explicitly allowed
This prevents every caller from needing to check if the location is
in the repo's .git directory. This also fixes a few places that we were
not doing this check before which caused various odd behaviors.
Also, add a debug message if we try to determine the status of a
location in the .git directory, before this patch it was possible
and would cause a warning.