git: Correctly update the view activatable
Don't just rediff the repository's contents when the focus is
returned to the window and update the view when the file
monitor emits a changed event for the location. Finally,
disable the file monitors when the window loses focus and
then do a full update when it returns. This prevents us
from doing pointless updates and allows us to react to git
commands made while the window didn't have focus.