Fix translation of plugins metainfo
It's not translated because INTLTOOL_XML_NOMERGE_RULE is used, but the
metainfo files require that the translations be merged into them. So
switch to normal INTLTOOL_XML_RULE.

Now, gedit-plugins was relying on INTLTOOL_XML_NOMERGE_RULE for its
GSettings translation support. This is pointless as intltool has had
GSettings support since 2012. Rely on it. This avoids the problem that
you can't use both the merge rule and the nomerge rule in the same
Makefile.am.

gedit-plugins should really not be using intltool at all anymore. It's
obsolete. Switching is not hard, but I'm not going to fix that today.

https://bugzilla.gnome.org/show_bug.cgi?id=788134